Move Datafetch to hook
